requirements.txt
pandas
requests
beautifulsoup4
openpyxl
pillow
pygame
